admission_api/src/issue_card/requirements.txt
web3==6.15.1
wheel
pillow==10.2.0